qlog: support for JSON-SEQ serialization format
To date, the qlog crate has support only the JSON serialization format.
In simple terms, that means a Trace object that contains an array of
Events. To avoid buffering, the qlog crate provides a streaming
serializer component for the non-streaming JSON format. Events are
appended to the end of a single line, written to a Writer. When
complete, the Trace object is finalized.

draft-ietf-quic-qlog-main-schema-01 added support for the JSON-SEQ
format, which is a true streaming format. Traces and Events are rendered
as individual records each prefixed by an ASCII Record Separator (<RS>,
0x1E), and each ending with an ASCII Line Feed character (\n, 0x0A).

This change adds support for JSON-SEQ to the core data model. The
streaming serializer has been updated to use the JSON-SEQ format.